Simple restructure to have all the databases in a more logical file location, furthermore, specific reading of database outputs are moved out of utils, and into more logical places - such as the database itself. blast_tools is also now a container for static methods for reading all things blast.
Removed all references to utils.py

minor unit testing foundation for databases.
