+++ commec/tests/test_dbs/nr.fasta
@@ -1,0 +1,4 @@
+> TEST_NR_01"NR01"|taxid=562
+SQSVSQSVSQSVSQSVSQSVSQSVSQSVSQSVSQSVSQSVSQ
+> TEST_NR_02"NR01_truncated"|taxid=562
+SQSVSQSVSQSVSQSVSQ
