+++ commec/tests/test_dbs/nt.fasta
@@ -1,0 +1,2 @@
+> NT_01"TEST01"|taxid=562
+AGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTCAGTC
